feat: Add horario_funcionamento and horario_fechamento columns to Restaurant model
This commit adds two new columns, `horario_funcionamento` and `horario_fechamento`, to the `Restaurant` model in the `restaurant_model.py` file. These columns are of type `DateTime` and have default values set to the current timestamp. They will be used to store the opening and closing times of the restaurant.